+++ src/grepros/__init__.py
@@ -11,8 +11,8 @@ Released under the BSD License.
 """
 ## @namespace grepros
 __title__        = "grepros"
-__version__      = "1.0.5.dev10"
-__version_info__ = (1, 0, 5, "dev10")
+__version__      = "1.0.5.dev11"
+__version_info__ = (1, 0, 5, "dev11")
 __version_date__ = "27.12.2023"
 
 from . library import *