+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog for package grepros
 
 1.1.0 (2023-12-27)
 -------------------
+* add support for splitting files in bag/HTML/MCAP/SQLite output
 * add api.to_duration()
 * do not use partial match for topic and typename filters without wildcards
 * support patterns yielding zero-length matches like "(?!not_this)"